refactor: reduce Scenario to bare minimum to ensure serialization
