+++ src/Xbehave.2.Execution.desktop/ScenarioOutlineRunner.cs
@@ -17,8 +17,9 @@ namespace Xbehave.Execution
         private static readonly object[] noArguments = new object[0];
 
         private readonly IMessageSink diagnosticMessageSink;
+        private readonly ScenarioRunnerFactory scenarioRunnerFactory;
         private readonly ExceptionAggregator cleanupAggregator = new ExceptionAggregator();
-        private readonly List<Scenario> scenarios = new List<Scenario>();
+        private readonly List<ScenarioRunner> scenarioRunners = new List<ScenarioRunner>();
         private Exception dataDiscoveryException;
 
         public ScenarioOutlineRunner(
@@ -41,6 +42,17 @@ namespace Xbehave.Execution
                 cancellationTokenSource)
         {
             this.diagnosticMessageSink = diagnosticMessageSink;
+            this.scenarioRunnerFactory = new ScenarioRunnerFactory(
+                this.TestCase,
+                this.DisplayName,
+                this.MessageBus,
+                this.TestClass,
+                this.ConstructorArguments,
+                this.TestMethod,
+                this.SkipReason,
+                this.BeforeAfterAttributes,
+                this.Aggregator,
+                this.CancellationTokenSource);
         }
 
         protected IMessageSink DiagnosticMessageSink
@@ -54,7 +66,7 @@ namespace Xbehave.Execution
 
             try
             {
-                var dataAttributes = TestCase.TestMethod.Method.GetCustomAttributes(typeof(DataAttribute)).ToList();
+                var dataAttributes = this.TestCase.TestMethod.Method.GetCustomAttributes(typeof(DataAttribute)).ToList();
                 foreach (var dataAttribute in dataAttributes)
                 {
                     var discovererAttribute = dataAttribute.GetCustomAttributes(typeof(DataDiscovererAttribute)).First();
@@ -63,31 +75,13 @@ namespace Xbehave.Execution
 
                     foreach (var dataRow in discoverer.GetData(dataAttribute, TestCase.TestMethod.Method))
                     {
-                        var scenario = new Scenario(
-                            this.TestCase,
-                            this.DisplayName,
-                            this.TestClass,
-                            this.TestMethod,
-                            dataRow,
-                            this.SkipReason,
-                            this.BeforeAfterAttributes);
-
-                        this.scenarios.Add(scenario);
+                        this.scenarioRunners.Add(this.scenarioRunnerFactory.Create(dataRow));
                     }
                 }
 
-                if (!this.scenarios.Any())
+                if (!this.scenarioRunners.Any())
                 {
-                    var scenario = new Scenario(
-                        this.TestCase,
-                        this.DisplayName,
-                        this.TestClass,
-                        this.TestMethod,
-                        noArguments,
-                        this.SkipReason,
-                        this.BeforeAfterAttributes);
-
-                    this.scenarios.Add(scenario);
+                    this.scenarioRunners.Add(this.scenarioRunnerFactory.Create(noArguments));
                 }
             }
             catch (Exception ex)
@@ -109,23 +103,18 @@ namespace Xbehave.Execution
             }
 
             var summary = new RunSummary();
-            foreach (var scenario in this.scenarios)
+            foreach (var scenarioRunner in this.scenarioRunners)
             {
-                summary.Aggregate(await scenario.RunAsync(
-                        this.diagnosticMessageSink,
-                        this.MessageBus,
-                        this.ConstructorArguments,
-                        new ExceptionAggregator(this.Aggregator),
-                        this.CancellationTokenSource));
+                summary.Aggregate(await scenarioRunner.RunAsync());
             }
 
             // Run the cleanup here so we can include cleanup time in the run summary,
             // but save any exceptions so we can surface them during the cleanup phase,
             // so they get properly reported as test case cleanup failures.
             var timer = new ExecutionTimer();
-            foreach (var scenario in this.scenarios)
+            foreach (var scenarioRunner in this.scenarioRunners)
             {
-                timer.Aggregate(() => this.cleanupAggregator.Run(() => scenario.Dispose()));
+                timer.Aggregate(() => this.cleanupAggregator.Run(() => scenarioRunner.Dispose()));
             }
 
             summary.Time += timer.Total;
